Hack to smooth out install name issues with CMake
So... CMake is a bit too clever when it comes to setting up loader paths. When
compiling poppler-qt4, it links against libpoppler-0.13.0.0.  When TeXworks
gets compiled, it links against libpoppler (which is a symlink). This causes
FIXUP_BUNDLE to fail or import two copies of the core poppler library when
packaging TeXworks.app.

This patch removes VERSION and SOVERSION information from the poppler libraries
which simplifies things by removing all dylib version symlinks. Having
versioned libraries is not so important for Mac packaging as the user will
never be upgrading the poppler libraries used by a specific TeXworks build.