<QtPDF> Play around with handling PageUp/PageDown events
This is a very dumb first cut at PageUp/PageDown handling. Will replace it with
something more sophisticated later and may delegate the actual interception of
KeyPress events to a parent widget.

Scroll events need to be captured so that the widget can track the current page
being viewed.

Appears the `item` method of `QGraphicsScene`/`QGraphicsView` returns graphics
items in an order that is the reverse of the one in which they were added. Each
object may possibly return a different order.  Should definitely place all
`PDFPageGraphicsItem` objects in a `QGraphicsItemGroup` --- possibly subclassed
so tha the order is explicitly handled.