<QtPDF> Track current page
The `PDFDocumentView` class now overloads `paintEvent` to keep track of which
page is currently being viewed.  The "current page" is defined as the first
graphics object in the list of objects occupying the upper half of the
viewport. If any other types of graphics objects are added, this should be
restricted to only objects which are of class `PDFPageGraphicsItem`.

All pointers to `PDFPageGraphicsItem` are now added to a `QList` as well as the
graphics scene. This provides an object that lists the pages in the order they
occur in the PDF document rather than the order that the graphics device
currently has them in.