<QtPDF> Move non-disposable code into `src`
The PDFDocumentView code will eventually be merged into the `src` subfolder of
another project. Everything else is a disposable testing application. This
change makes it easy to split the non-disposable code into a subtree once it is
ready.

Makefile is getting a little tricky, but I am going to keep using it.