<QtPDF> Implement first cut at a tile map
The tile map is a matrix of QRect objects that is calculated for each zoom
level such that each tile is 1024 pixels square (a pixmap of ~4MB assuming 4
bytes of data per pixel).

Currently no tiles are rendered, their outlines are merely painted over the
document scene.

Lesson learned:

It is not necessary to reset the Painter transform to the identity matrix
before painting page objects---just set the scale components back to 1. This
preserves the translation elements which removes the need to do all
transformations up front, such as casting pixmap origins.