<QtPDF> Enhance interaction with search results
Signals and Slots are in place that allow the forward, back and clear buttons
on the `SearchLineEdit` objects to interact with searching functions in
`PDFDocumentView` objects.

Management of the list of `QGraphicsRectItem` objects representing search
results has been transferred from `PDFDocumentScene` to `PDFDocumentView`.  Not
completely happy with this setup, but it makes the most sense right now.