Fix default saving location (issue 609)
When saving to a file that did not exist, QFileInfo::canonicalPath() returned "/". But in case of symlinks, the canonical path is not what we want, anyway. So this was replaced now by QFileInfo::absolutePath(), which works as expected.

In addition, the home directory was set as fallback if saveDialogDir was set to a non-existent directory (e.g., because it was unmounted or disconnected).