Replace svn revision info by git commit info in build descriptions
Where applicable, the (abbreviated) git commit hash together with the commit date is given. The hash enables to track down the actual code version, whereas the date facilitates to check if a particular version is "newer" than another (e.g., for users who wonder if they should update or not).

Note that, by design of a distributed version control system such as git, there is no (reliable) way to assign incremental numbers (such as svn revision numbers) to commits.