Fix crash when syncing from a pdf to a (closed) tex that is linked to said pdf (issue #670)
In TWSyncTeXSynchronizer::_syncFromPDFFine(), the tex file is opened. This triggers TeXDocument::openPdfIfAvailable() to open the pdf (from which the synctex trigger originated). Previously, TeXDocument::openPdfIfAvailable() caused a complete reload of the pdf. This entailed destroying the TWSyncTeXSynchronizer object which was still processing the syncing request.
However, forcing a reload should not be necessary as the PDFDocument is guarded by a QFileSystemWatcher anyway.