Fix and simplify selection of OS-specific code
The fix is related to the fact that `Q_OS_UNIX` is also defined on OS X,
so in order to properly check for Linux and other non-OS-X UNIXes one
has to check for `Q_OS_UNIX` and absence of `Q_OS_DARWIN` (a synonym for
`Q_OS_MAC`, but `Q_OS_MAC` is only defined for Qt >= 4.6).

The simplification is about eliminating double checks for both `Q_WS_*`
and `Q_OS_*` (the latter were added for Qt5 compatibility). The idea is
to reduce all checks to `Q_OS_*`, more precisely:

* OS X: `defined(Q_OS_DARWIN)`
* Windows: `defined(Q_OS_WIN)`
* Linux et al.: `defined(Q_OS_UNIX) && !defined(Q_OS_DARWIN)`

Additional notes:

* The previously used check for both `Q_OS_UNIX` and `Q_OS_LINUX` was
  redundant as `Q_OS_UNIX` is basically defined as `!Q_OS_WIN` (checked
  with Qt 4.8.6 and 5.4.1, minus unsupported platforms like MS-DOS and
  OS/2).

* The check for `Q_OS_DARWIN` can be replaced with `Q_OS_MAC` as soon as
  the Qt requirement is raised to >= 4.6. This can be done with some
  simple search and replace and should make the code easier to
  understand for future contributors.