Ensure 'trans/TeXworks_trans.pro' is always sorted
The file 'trans/TeXworks_trans.pro' is auto-generated, but kept under
version control to ease the work of translators. Make sure that the
order of the entries in the generated file is not affected by globbing
order which depends on the file and operating system. This should
prevent future spurious changes to this file.

Also updates 'trans/TeXworks_trans.pro' according to the above logic.