Avoid updating the PDF preview during typesetting
During typesetting, the output pdf naturally is changed. Reloading it while it is being changed is not advisable as we might end up with a broken pdf or at least an out-of-date version. For long TeX runs, reloading could actually be triggered multiple times.
To avoid such situations, watching for changes to the pdf file on disk by the previewer is now disabled while a typesetting process is running.