Fix crash when sync-clicking in the source during typesetting
When syncing, a highlight graphics item is created. The PDFDocument instance keeps a pointer to that so it can be hidden after kPDFHighlightDuration. When typesetting finishes, the pdf is reloaded, all old pages are destroyed, and, in the process, all subordinate items are deleted as well. Thus, PDFDocument then holds a pointer to an object that no longer exists, causing a segfault.
This problem is now remedied by invalidating the pointer when the document is changed.