<QtPDF> Implement PDFDocumentView::goPrevViewRect() and PDFDocumentView::canGoPrevViewRects()
Whenever a link is clicked, the old view rect is stored (as PDFDestination) on the stack PDFDocumentView::_oldViewRects. By invoking PDFDocumentView::goPrevViewRect(), the user can quickly get back to those previous views. Typical use cases include: getting back to the ToC after clicking on an item, getting back to the index after clicking on an item, getting back to the text after clicking on a cross-reference (e.g., to an image, bibliography, etc.).

In the PDFViewer demo app, PDFDocumentView::goPrevViewRect() is connected to the Alt+Left shortcut.