<QtPDF> Speed up text select by using simple rects instead of complex regions
To avoid holes in occuring where two boxes intersect, previously a large and complex highlight path was constructed as union of all box rects. For larger selection areas (~half a page), this became extremely slow.
Now, we simply add a list of rects but with a properly set winding rule to avoid holes.