<QtPDF> Fix Page::selectedText() for not-enclosing selections (e.g. SyncTeX)
In 3d4e34, the behavior of selectedText() was changed to return only characters whose char boxes were entirely contained in the selection polygon instead of just intersected by it. The rationale behind that was that if the selection polygon is made of (entire) char boxes, and some char boxes overlap, using all chars from intersecting boxes would include characters that are not actually selected.
However, for other uses - such as SyncTeX - the selection is not made up of PDF boxes, but rather other shapes (e.g. SyncTeX boxes) that may not full enclose the PDF char boxes. Thus, selectedText() would fail to return the proper text if it's looking only for fully enclosed boxes.

In this revision, the new parameter "onlyFullyEnclosed" (default: false, i.e. pre-3d4e34 behavior) is introduced to be able to distinguish those two cases.