<QtPDF> Remember which page(s) were shown in PDFDocumentScene across (possibly broken) reloads
Especially in single-page non-continuous mode, it is important that the correct page is shown after a reload (in continuous mode, all pages are always shown; in non-continuous mode, all pages are stacked on top of each other and only the one, correct page is set to "visible"). Previously, we determined the visibility of pages before reloading and then tried to recreate that state after reloading. However, in cases where the PDF went through several stages of being broken (e.g., because TeX was not finished with it yet), the first reload deleted all pages, so subsequent reloads had no visibility data to work with. Thus, QtPDF reverted back to page 1.
Now, we track the currently shown page using PDFDocumentScene::_shownPageIdx which does not get reset when the PDF is broken.

Fixes #718