<QtPDF> Set the reload timer delay to 500 ms
The reload timer delay is used to try to avoid reloading PDFs while they are changing. E.g., when TeX writes to the PDF, it does so page-by-page, so the file on the harddisk changes multiple times. If two such changes happen within the reload timer delay, the first one is disregarded.
Previously, the reload timer delay was 100 ms, which seems to be too short for complex TeX documents (e.g., with large figures or a lot of processing per-page).

Affects #718