<QtPDF> Use old images as rendering placeholers even after document reload (issue #735)
Instead of using the "rendering page" as placeholder, keep PDF tile images in the cache even when the document is reloaded, mark them as outdated (so the program knows it needs to trigger a rerender), and use them as placeholder. This avoids the flickering display of "rendering page" for a split-second which impedes checking small changes across typesetting runs.