Implement NonblockingSyntaxHighlighter (fixes #771, affects #469)
NonblockingSyntaxHighlighter is a reimplementation (and replacement) of
QSyntaxHighlighter, with the aim of keeping the UI responsive even
during lengthy highlighting tasks (e.g., long documents). It achieves
this goal by queuing highlighting tasks and executing them for a certain
time (a few milliseconds) before returning control to the main event loop
for processing other events. Once the other events are processed, the
highlighter resumes its work for a few milliseconds and so on until the
whole documented is highlighted.

Inspired by http://enki-editor.org/2014/08/22/Syntax_highlighting.html