Add syntax highlighting for Lua code
As Lua has a short list of 'real' keywords these can be picked out
separately from everything else. Notice that as we are doing line-by-line
highlighting there is no chance of covering Lua's --[[ ... --]] syntax
for comments.

Possibly the string-matching regexes can be made more efficient: may
go with the request for the Qt5 PCRE regex engine to be enabled here.