Fix synchronization when searching in a PDF (fixes #693)
Previously, not the whole search string was highlighted in the TeX
source. Instead, only one character (typically the first character of
the search string) was highlighted.
Now, the whole search string is highlighted.

Caveats: Due to the way spaces are handled in the PDF previewer when
extracting text, leading and trailing spaces are still not highlighted
properly.