<QtPDF> Remove "-platform offscreen" from tests
While tests do start with "-platform offscreen", those that require
system fonts fail with a "QFontDatabase: Cannot find font directory"
error. This is due to the fact that QFontDatabase relies on the platform
integration for finding the system font directory, and the "offscreen"
platform integration apparently reports an incorrect/unusable one.