Make window titles unique (fixes #853)
Each window's title (and the corresponding entry in the Window menu) now
uniquely identify the open file. I.e., as long as no two open files
share the same filename, only the filename is displayed (as before). If
two or more open files share the same filename, parts of the directory
tree are displayed to uniquely identify the files.
This also means that window titles can change dynamically, e.g. when
some file is opened, closed, or saved under a new name.