Migrate custom Poppler GlobalParams handling to QtPDF
Poppler requires additional encoding data (poppler-data) to handle
certain files, mostly in CJK languages. While poppler-data is typically
installed system-wide when TeXworks is installed, this is not the case
on Windows and Mac OS X. To accomodate those platforms, poppler-data is
bundled with TeXworks. For poppler to find the data, its GlobalParams
need to know about their location. On Windows, this works out-of-the-box
now, but on Mac OS X this still needs to be patched/hacked.

TeXworks should not need to know about any such implementation details of
QtPDF.