<QtPDF> Make `UpdateTranslations` a separate build command not run by default
Updating files in the source tree is bad if one wants to have several
different builds (e.g. with different parameters) coexisting. Besides,
the .pro helper files created by this command are not used in the build
process so there is no need to update them on every build.

Instead, run UpdateTranslations manually when you need to.