Don't reload a file if its content hasn't changed (#901)
Even if the file modification date has changed, if the content is still
the same, there is no point in re-setting the text (thereby triggering a
bunch of side-effects such as destroying the undo/redo stack). Such a
situation can occur, e.g., if modification times are inaccurate, messed
with by something like a cloud synchronization process, or if the file
is simply `touch`ed.