Initial support for QJSEngine-based scripting (#893)
QtScript has been deprecated for a while in favor of the new QJSEngine.
Also, QtScript will be removed in Qt6.
For now, QJSEngine supports the .es and .js extensions, although
QtScript takes precedence for .js scripts for backwards compatibility.
Note that QJSEngine provdes a stack trace for errors, but no debugger
(yet; QTBUG-55905)