Set objectName for typesetting tools so they can be assigned a shortcut (fixes #906)
The objectName (which can be assigned a shortcut via shortcuts.ini - see
the manual) is set to actionTypesetTool#TOOLNAME#, where #TOOLNAME# is
the name as shown in the menu/dropdown list. Examples include
"actionTypesetToolpdfLaTeX" (for "pdfLaTeX"), "actionTypesetToolMy tool"
(for "My tool"), and "actionTypesetToolA%3DB" (for "A=B"). Note that
some characters (such as "=" and ";" are reserved in ini files and must
be escaped by "%XX", where XX is the corresponding hexadecimal ASCII code).