Fix preview filename construction (issue #908)
Especially for files in Windows root directories (e.g., C:\, D:\, ...),
the preview filename was constructed incorrectly (C:\\a.pdf instead of
C:\a.pdf). This resulted in the failure to recognize updates to the
preview file (as the filename seemed to have changed) and consequently
in the pdf display not updating correctly after typesetting.