Fix PDF preview scale context menu (fixes #953)
Previously, the context menu was only constructed once in the program
(as static variable) on first use. This meant that it was always tied to
the window in which it was first used. Subsequent uses in a separate
window resulted in changing the zoom in the first window. Even worse, if
the first window was closed (destroyed), the menu referenced a dangling
pointer, resulting in a crash.
Now, the context menu is constructed once per PDF preview window (and
stored as a private property). It is destroyed automatically when its
parent QObject (the PDF preview window) is destroyed.