Prevent concurrent typesetting in multi-file sources
For any given (root) document, Only one typesetting process should run
at the same time (otherwise, the auxiliary and output files would be
messed up).
With this commit, Tw detects if a typesetting for the same root document
was already started from another Tw window. If so, the typesetting
button takes the user to the window in which the process is currently
running.

Fixes #911