<QtPDF> Start a new search when the search flags have changed
Even if the search string is the same, the actual search depends on
flags such as case sensitivity. Some flags (e.g. concerning the search
direction) may not alter the results themselves (but rather their
ordering) and could potentially be implemented cleverly, but for now
always starting with a clean slate seems to be the safer option.