<QtPDF> Fix range of guidline dialog position field
Set range to -99999.99 .. 99999.99 (instead of the default 0 .. 99.99).
This should be sufficient even for large page sizes (meters), and for
small units (such as bp) (10 m ~ 28000 bp). Also works for guidelines
off the left (or top) of the page, i.e., negative coordinates.

Reported by Philip Taylor on the mailing list
[https://tug.org/pipermail/texworks/2023q4/007457.html]